requirements.txt
browser-use==0.5.4
python-dotenv==1.1.0
langchain-openai==0.3.27
playwright==1.53.0
flask==3.0.3
